refactor(sharepoint): extract pure parsing helpers into parsing.py
First safe step toward decomposing the content-service god-object: move the
stateless parsing/content helpers (canvas extraction, text sanitisation, facet
detection, unsupported-file detection, safe_int, require_text) into a dedicated
parsing module with public names. Call sites and tests import from parsing; no
behaviour change.
